% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ph_kde.R
\docType{class}
\name{TDAKernel-class}
\alias{TDAKernel-class}
\alias{TDAKernel}
\title{Create a persistence diagram kernel density.}
\description{
A reference class to represent the persistence diagram kernel from Maroulas, Mike, and Oballe 2019.
It allows sampling and approximate evaluation.
}
\section{Fields}{

\describe{
\item{\code{center}}{A diagram created from the TDA package. This is the center of the kernel.}

\item{\code{sigma}}{A numeric. This is the bandwidth of the kernel.}

\item{\code{lower.card.probs}}{A vector of probabilities for to serve as the cardinality distribution of the lower portion of the diagram. The probability at index i is the probability of observing i points from the lower diagram.}

\item{\code{upper}}{A matrix whose rows are points in the upper portion of the kernel in birth-persistence coordinates. Created automatically on initialization.}

\item{\code{lower}}{A matrix whose rows are points in the lower portion of the kernel in birth-persistence coordinates. Created automatically on initialization.}

\item{\code{upper.probs}}{A vector of probabilities of appearances for each point in the upper diagram. The ith entry is the probability of appearance corresponding to row i of upper. Computed automatically on intialization.}
}}

\section{Methods}{

\describe{
\item{\code{eval(D)}}{Approximate evaluation of a diagram D (a diagram from the TDA package) in the kernel. Evaluation is the sum over highest probability matchings for all possible upper diagram cardinalities.}

\item{\code{initialize(center, sigma, lower.card.probs, fast.eval = TRUE)}}{Creates the TDAKernel object. Requires input of center, sigma, and lower.card.probs. User can also designate fast.eval = FALSE to assume all upper.probs are 1.}

\item{\code{sample(n)}}{Draw n samples from the kernel.}
}}

